% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{fct_max}
\alias{fct_max}
\title{Find max ordinal level in vector}
\usage{
fct_max(x)
}
\arguments{
\item{x}{A vector representing ordinal values}
}
\value{
A character vector representing the max/min ordinal value in \code{x}
}
\description{
Returns the maximum ordinal level present in the vector
}
\details{
Factors generally are numerical values representing specific strings that
are expected to be repeated in a vector (see iris$Species for instance). Ordinal factors
are factors that have an ordering imposed by the levels specified in the factor (see
\code{\link[=levels]{levels()}}).

This function determines, for a vector, what the smallest (or largest) ordinal factor
is. In a vector it is not required that all ordinal values are present. This function
will return the factor level of the smallest value (or largest value) in the vector.
}
\examples{
\dontrun{
find_level(iris$Species)
# setosa
find_level(iris$Species, f = max)
# virginica
}
}
\seealso{
fct_min(), fct_max()
}
