% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise.fisher.test}
\alias{pairwise.fisher.test}
\title{Tabulate pairwise comparisons for all levels}
\usage{
pairwise.fisher.test(g1, g2, p.adjust.method = p.adjust.methods)
}
\arguments{
\item{p.adjust.method}{}
}
\description{
Tabulate pairwise comparisons for all levels
}
